Address busbey feedback:
* Fix up metrics source interface hierarchy
* Docs for test replication endpoint
* Push down metrics updating to mutation of the buffer usage variable
  (rather than on a timer)
