+++ hbase-hadoop-compat/src/main/java/org/apache/hadoop/hbase/replication/regionserver/MetricsReplicationGlobalSourceSourceImpl.java
@@ -1,4 +1,4 @@
-/*
+/**
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -15,6 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.apache.hadoop.hbase.replication.regionserver;
 
 import org.apache.hadoop.metrics2.lib.MutableFastCounter;
@@ -23,22 +24,10 @@ import org.apache.hadoop.metrics2.lib.MutableHistogram;
 import org.apache.yetus.audience.InterfaceAudience;
 
 @InterfaceAudience.Private
-public class MetricsReplicationSourceSourceImpl implements MetricsReplicationSourceSource {
+public class MetricsReplicationGlobalSourceSourceImpl implements MetricsReplicationGlobalSourceSource {
+  private static final String KEY_PREFIX = "source.";
 
   private final MetricsReplicationSourceImpl rms;
-  private final String id;
-  private final String sizeOfLogQueueKey;
-  private final String ageOfLastShippedOpKey;
-  private final String logReadInEditsKey;
-  private final String logEditsFilteredKey;
-  private final String shippedBatchesKey;
-  private final String shippedOpsKey;
-  private String keyPrefix;
-
-  private final String shippedBytesKey;
-  private final String logReadInBytesKey;
-  private final String shippedHFilesKey;
-  private final String sizeOfHFileRefsQueueKey;
 
   private final MutableHistogram ageOfLastShippedOpHist;
   private final MutableGaugeLong sizeOfLogQueueGauge;
@@ -50,14 +39,6 @@ public class MetricsReplicationSourceSourceImpl implements MetricsReplicationSou
   private final MutableFastCounter logReadInBytesCounter;
   private final MutableFastCounter shippedHFilesCounter;
   private final MutableGaugeLong sizeOfHFileRefsQueueGauge;
-
-  private final String unknownFileLengthKey;
-  private final String uncleanlyClosedKey;
-  private final String uncleanlySkippedBytesKey;
-  private final String restartedKey;
-  private final String repeatedBytesKey;
-  private final String completedLogsKey;
-  private final String completedRecoveryKey;
   private final MutableFastCounter unknownFileLengthForClosedWAL;
   private final MutableFastCounter uncleanlyClosedWAL;
   private final MutableFastCounter uncleanlyClosedSkippedBytes;
@@ -65,62 +46,47 @@ public class MetricsReplicationSourceSourceImpl implements MetricsReplicationSou
   private final MutableFastCounter repeatedFileBytes;
   private final MutableFastCounter completedWAL;
   private final MutableFastCounter completedRecoveryQueue;
+  private final MutableFastCounter failedRecoveryQueue;
+  private final MutableGaugeLong walReaderBufferUsageBytes;
 
-  public MetricsReplicationSourceSourceImpl(MetricsReplicationSourceImpl rms, String id) {
+  public MetricsReplicationGlobalSourceSourceImpl(MetricsReplicationSourceImpl rms) {
     this.rms = rms;
-    this.id = id;
-    this.keyPrefix = "source." + this.id + ".";
 
-    ageOfLastShippedOpKey = this.keyPrefix + "ageOfLastShippedOp";
-    ageOfLastShippedOpHist = rms.getMetricsRegistry().getHistogram(ageOfLastShippedOpKey);
+    ageOfLastShippedOpHist = rms.getMetricsRegistry().getHistogram(SOURCE_AGE_OF_LAST_SHIPPED_OP);
 
-    sizeOfLogQueueKey = this.keyPrefix + "sizeOfLogQueue";
-    sizeOfLogQueueGauge = rms.getMetricsRegistry().getGauge(sizeOfLogQueueKey, 0L);
+    sizeOfLogQueueGauge = rms.getMetricsRegistry().getGauge(SOURCE_SIZE_OF_LOG_QUEUE, 0L);
 
-    shippedBatchesKey = this.keyPrefix + "shippedBatches";
-    shippedBatchesCounter = rms.getMetricsRegistry().getCounter(shippedBatchesKey, 0L);
+    shippedBatchesCounter = rms.getMetricsRegistry().getCounter(SOURCE_SHIPPED_BATCHES, 0L);
 
-    shippedOpsKey = this.keyPrefix + "shippedOps";
-    shippedOpsCounter = rms.getMetricsRegistry().getCounter(shippedOpsKey, 0L);
+    shippedOpsCounter = rms.getMetricsRegistry().getCounter(SOURCE_SHIPPED_OPS, 0L);
 
-    shippedBytesKey = this.keyPrefix + "shippedBytes";
-    shippedBytesCounter = rms.getMetricsRegistry().getCounter(shippedBytesKey, 0L);
+    shippedBytesCounter = rms.getMetricsRegistry().getCounter(SOURCE_SHIPPED_BYTES, 0L);
 
-    logReadInBytesKey = this.keyPrefix + "logReadInBytes";
-    logReadInBytesCounter = rms.getMetricsRegistry().getCounter(logReadInBytesKey, 0L);
+    logReadInBytesCounter = rms.getMetricsRegistry().getCounter(SOURCE_LOG_READ_IN_BYTES, 0L);
 
-    logReadInEditsKey = this.keyPrefix + "logEditsRead";
-    logReadInEditsCounter = rms.getMetricsRegistry().getCounter(logReadInEditsKey, 0L);
+    logReadInEditsCounter = rms.getMetricsRegistry().getCounter(SOURCE_LOG_READ_IN_EDITS, 0L);
 
-    logEditsFilteredKey = this.keyPrefix + "logEditsFiltered";
-    walEditsFilteredCounter = rms.getMetricsRegistry().getCounter(logEditsFilteredKey, 0L);
+    walEditsFilteredCounter = rms.getMetricsRegistry().getCounter(SOURCE_LOG_EDITS_FILTERED, 0L);
 
-    shippedHFilesKey = this.keyPrefix + "shippedHFiles";
-    shippedHFilesCounter = rms.getMetricsRegistry().getCounter(shippedHFilesKey, 0L);
+    shippedHFilesCounter = rms.getMetricsRegistry().getCounter(SOURCE_SHIPPED_HFILES, 0L);
 
-    sizeOfHFileRefsQueueKey = this.keyPrefix + "sizeOfHFileRefsQueue";
-    sizeOfHFileRefsQueueGauge = rms.getMetricsRegistry().getGauge(sizeOfHFileRefsQueueKey, 0L);
+    sizeOfHFileRefsQueueGauge =
+        rms.getMetricsRegistry().getGauge(SOURCE_SIZE_OF_HFILE_REFS_QUEUE, 0L);
 
-    unknownFileLengthKey = this.keyPrefix + "closedLogsWithUnknownFileLength";
-    unknownFileLengthForClosedWAL = rms.getMetricsRegistry().getCounter(unknownFileLengthKey, 0L);
+    unknownFileLengthForClosedWAL = rms.getMetricsRegistry()
+            .getCounter(SOURCE_CLOSED_LOGS_WITH_UNKNOWN_LENGTH, 0L);
+    uncleanlyClosedWAL = rms.getMetricsRegistry().getCounter(SOURCE_UNCLEANLY_CLOSED_LOGS, 0L);
+    uncleanlyClosedSkippedBytes = rms.getMetricsRegistry()
+            .getCounter(SOURCE_UNCLEANLY_CLOSED_IGNORED_IN_BYTES, 0L);
+    restartWALReading = rms.getMetricsRegistry().getCounter(SOURCE_RESTARTED_LOG_READING, 0L);
+    repeatedFileBytes = rms.getMetricsRegistry().getCounter(SOURCE_REPEATED_LOG_FILE_BYTES, 0L);
+    completedWAL = rms.getMetricsRegistry().getCounter(SOURCE_COMPLETED_LOGS, 0L);
+    completedRecoveryQueue = rms.getMetricsRegistry()
+            .getCounter(SOURCE_COMPLETED_RECOVERY_QUEUES, 0L);
+    failedRecoveryQueue = rms.getMetricsRegistry()
+            .getCounter(SOURCE_FAILED_RECOVERY_QUEUES, 0L);
 
-    uncleanlyClosedKey = this.keyPrefix + "uncleanlyClosedLogs";
-    uncleanlyClosedWAL = rms.getMetricsRegistry().getCounter(uncleanlyClosedKey, 0L);
-
-    uncleanlySkippedBytesKey = this.keyPrefix + "ignoredUncleanlyClosedLogContentsInBytes";
-    uncleanlyClosedSkippedBytes = rms.getMetricsRegistry().getCounter(uncleanlySkippedBytesKey, 0L);
-
-    restartedKey = this.keyPrefix + "restartedLogReading";
-    restartWALReading = rms.getMetricsRegistry().getCounter(restartedKey, 0L);
-
-    repeatedBytesKey = this.keyPrefix + "repeatedLogFileBytes";
-    repeatedFileBytes = rms.getMetricsRegistry().getCounter(repeatedBytesKey, 0L);
-
-    completedLogsKey = this.keyPrefix + "completedLogs";
-    completedWAL = rms.getMetricsRegistry().getCounter(completedLogsKey, 0L);
-
-    completedRecoveryKey = this.keyPrefix + "completedRecoverQueues";
-    completedRecoveryQueue = rms.getMetricsRegistry().getCounter(completedRecoveryKey, 0L);
+    walReaderBufferUsageBytes = rms.getMetricsRegistry().getGauge(SOURCE_WAL_READER_EDITS_BUFFER, 0L);
   }
 
   @Override public void setLastShippedAge(long age) {
@@ -160,29 +126,6 @@ public class MetricsReplicationSourceSourceImpl implements MetricsReplicationSou
   }
 
   @Override public void clear() {
-    rms.removeMetric(ageOfLastShippedOpKey);
-
-    rms.removeMetric(sizeOfLogQueueKey);
-
-    rms.removeMetric(shippedBatchesKey);
-    rms.removeMetric(shippedOpsKey);
-    rms.removeMetric(shippedBytesKey);
-
-    rms.removeMetric(logReadInBytesKey);
-    rms.removeMetric(logReadInEditsKey);
-
-    rms.removeMetric(logEditsFilteredKey);
-
-    rms.removeMetric(shippedHFilesKey);
-    rms.removeMetric(sizeOfHFileRefsQueueKey);
-
-    rms.removeMetric(unknownFileLengthKey);
-    rms.removeMetric(uncleanlyClosedKey);
-    rms.removeMetric(uncleanlySkippedBytesKey);
-    rms.removeMetric(restartedKey);
-    rms.removeMetric(repeatedBytesKey);
-    rms.removeMetric(completedLogsKey);
-    rms.removeMetric(completedRecoveryKey);
   }
 
   @Override
@@ -190,8 +133,7 @@ public class MetricsReplicationSourceSourceImpl implements MetricsReplicationSou
     return ageOfLastShippedOpHist.getMax();
   }
 
-  @Override
-  public void incrHFilesShipped(long hfiles) {
+  @Override public void incrHFilesShipped(long hfiles) {
     shippedHFilesCounter.incr(hfiles);
   }
 
@@ -214,40 +156,34 @@ public class MetricsReplicationSourceSourceImpl implements MetricsReplicationSou
   public void incrUnknownFileLengthForClosedWAL() {
     unknownFileLengthForClosedWAL.incr(1L);
   }
-
   @Override
   public void incrUncleanlyClosedWALs() {
     uncleanlyClosedWAL.incr(1L);
   }
-
   @Override
   public void incrBytesSkippedInUncleanlyClosedWALs(final long bytes) {
     uncleanlyClosedSkippedBytes.incr(bytes);
   }
-
   @Override
   public void incrRestartedWALReading() {
     restartWALReading.incr(1L);
   }
-
   @Override
   public void incrRepeatedFileBytes(final long bytes) {
     repeatedFileBytes.incr(bytes);
   }
-
   @Override
   public void incrCompletedWAL() {
     completedWAL.incr(1L);
   }
-
   @Override
   public void incrCompletedRecoveryQueue() {
     completedRecoveryQueue.incr(1L);
   }
-
   @Override
-  public void incrFailedRecoveryQueue() {/*no op*/}
-
+  public void incrFailedRecoveryQueue() {
+    failedRecoveryQueue.incr(1L);
+  }
   @Override
   public void init() {
     rms.init();
@@ -255,32 +191,32 @@ public class MetricsReplicationSourceSourceImpl implements MetricsReplicationSou
 
   @Override
   public void setGauge(String gaugeName, long value) {
-    rms.setGauge(this.keyPrefix + gaugeName, value);
+    rms.setGauge(KEY_PREFIX + gaugeName, value);
   }
 
   @Override
   public void incGauge(String gaugeName, long delta) {
-    rms.incGauge(this.keyPrefix + gaugeName, delta);
+    rms.incGauge(KEY_PREFIX + gaugeName, delta);
   }
 
   @Override
   public void decGauge(String gaugeName, long delta) {
-    rms.decGauge(this.keyPrefix + gaugeName, delta);
+    rms.decGauge(KEY_PREFIX + gaugeName, delta);
   }
 
   @Override
   public void removeMetric(String key) {
-    rms.removeMetric(this.keyPrefix + key);
+    rms.removeMetric(KEY_PREFIX + key);
   }
 
   @Override
   public void incCounters(String counterName, long delta) {
-    rms.incCounters(this.keyPrefix + counterName, delta);
+    rms.incCounters(KEY_PREFIX + counterName, delta);
   }
 
   @Override
   public void updateHistogram(String name, long value) {
-    rms.updateHistogram(this.keyPrefix + name, value);
+    rms.updateHistogram(KEY_PREFIX + name, value);
   }
 
   @Override
@@ -303,15 +239,28 @@ public class MetricsReplicationSourceSourceImpl implements MetricsReplicationSou
     return rms.getMetricsName();
   }
 
-  @Override public long getWALEditsRead() {
+  @Override
+  public long getWALEditsRead() {
     return this.logReadInEditsCounter.value();
   }
 
-  @Override public long getShippedOps() {
+  @Override
+  public long getShippedOps() {
     return this.shippedOpsCounter.value();
   }
 
-  @Override public long getEditsFiltered() {
+  @Override
+  public long getEditsFiltered() {
     return this.walEditsFilteredCounter.value();
+  }
+
+  @Override
+  public void setWALReaderEditsBufferBytes(long usage) {
+    this.walReaderBufferUsageBytes.set(usage);
+  }
+
+  @Override
+  public long getWALReaderEditsBufferBytes() {
+    return this.walReaderBufferUsageBytes.value();
   }
 }
