+++ hbase-server/src/main/java/org/apache/hadoop/hbase/replication/regionserver/Replication.java
@@ -77,7 +77,7 @@ public class Replication implements ReplicationSourceService, ReplicationSinkSer
   private int statsThreadPeriod;
   // ReplicationLoad to access replication metrics
   private ReplicationLoad replicationLoad;
-  private MetricsReplicationSourceSource globalMetricsSource;
+  private MetricsReplicationGlobalSourceSource globalMetricsSource;
 
   private PeerProcedureHandler peerProcedureHandler;
 
@@ -126,10 +126,12 @@ public class Replication implements ReplicationSourceService, ReplicationSinkSer
       throw new IOException("Could not read cluster id", ke);
     }
     SyncReplicationPeerMappingManager mapping = new SyncReplicationPeerMappingManager();
+    this.globalMetricsSource = CompatibilitySingletonFactory.getInstance(MetricsReplicationSourceFactory.class)
+        .getGlobalSource();
     this.replicationManager = new ReplicationSourceManager(queueStorage, replicationPeers,
         replicationTracker, conf, this.server, fs, logDir, oldLogDir, clusterId,
         walProvider != null ? walProvider.getWALFileLengthProvider() : p -> OptionalLong.empty(),
-        mapping);
+        mapping, globalMetricsSource);
     this.syncReplicationPeerInfoProvider =
         new SyncReplicationPeerInfoProviderImpl(replicationPeers, mapping);
     PeerActionListener peerActionListener = PeerActionListener.DUMMY;
@@ -152,8 +154,6 @@ public class Replication implements ReplicationSourceService, ReplicationSinkSer
               p.getSyncReplicationState(), p.getNewSyncReplicationState(), 0));
       }
     }
-    this.globalMetricsSource = CompatibilitySingletonFactory.getInstance(MetricsReplicationSourceFactory.class)
-        .getGlobalSource();
     this.statsThreadPeriod =
         this.conf.getInt("replication.stats.thread.period.seconds", 5 * 60);
     LOG.debug("Replication stats-in-log period={} seconds",  this.statsThreadPeriod);
@@ -218,7 +218,7 @@ public class Replication implements ReplicationSourceService, ReplicationSinkSer
     this.replicationManager.init();
     this.replicationSink = new ReplicationSink(this.conf);
     this.scheduleThreadPool.scheduleAtFixedRate(
-      new ReplicationStatisticsTask(this.replicationSink, this.replicationManager, this.globalMetricsSource),
+      new ReplicationStatisticsTask(this.replicationSink, this.replicationManager),
       statsThreadPeriod, statsThreadPeriod, TimeUnit.SECONDS);
     LOG.info("{} started", this.server.toString());
   }
@@ -248,22 +248,17 @@ public class Replication implements ReplicationSourceService, ReplicationSinkSer
 
     private final ReplicationSink replicationSink;
     private final ReplicationSourceManager replicationManager;
-    private final MetricsReplicationSourceSource globalMetricsSource;
 
     public ReplicationStatisticsTask(ReplicationSink replicationSink,
-        ReplicationSourceManager replicationManager, MetricsReplicationSourceSource globalMetricsSource) {
+        ReplicationSourceManager replicationManager) {
       this.replicationManager = replicationManager;
       this.replicationSink = replicationSink;
-      this.globalMetricsSource = globalMetricsSource;
     }
 
     @Override
     public void run() {
       printStats(this.replicationManager.getStats());
       printStats(this.replicationSink.getStats());
-
-      // Report how much data we've read off disk which is pending replication, across all sources
-      globalMetricsSource.setWALReaderEditsBufferBytes(replicationManager.getTotalBufferUsed().get());
     }
 
     private void printStats(String stats) {
